Rendering the cube!
